Rename to resolve-dvs
